feat: add MPS support, raw-state-dict checkpoint loading, and single-spectrum inference script
- spec2mol_main.py: detect CUDA/MPS/CPU automatically; handle raw state dict
  checkpoints (Zenodo) separately from full PL checkpoints via _is_pl_checkpoint()
- distributions.py: cast prob tensor to float32 before moving to MPS device
  to avoid float64 unsupported error
- configs: switch default dataset to msg_mini, activate MSG large model dims
  (encoder_hidden_dim=512, magma_modulo=2048), disable wandb, set test_only path
- configs/dataset/msg_mini.yaml: new mini split config for quick inference tests
- scripts/spec2mol_infer_one.py: standalone single-spectrum inference script
  with CUDA > MPS > CPU auto-detection, progress display, and tabular output

+++ src/diffusion/distributions.py
@@ -24,7 +24,7 @@ class DistributionNodes:
 
     def log_prob(self, batch_n_nodes):
         assert len(batch_n_nodes.size()) == 1
-        p = self.prob.to(batch_n_nodes.device)
+        p = self.prob.float().to(batch_n_nodes.device)
 
         probas = p[batch_n_nodes]
         log_p = torch.log(probas + 1e-6)
